TST/REF: collect tests by method (#37372)
* TST: cln/parametrize

* TST: collect tests by method

* TST/REF: collect tests by method

* lint fixup
