+++ pandas/tests/series/test_reductions.py
@@ -2,7 +2,8 @@ import numpy as np
 import pytest
 
 import pandas as pd
-from pandas import Series
+from pandas import MultiIndex, Series
+import pandas._testing as tm
 
 
 def test_reductions_td64_with_nat():
@@ -44,6 +45,14 @@ def test_prod_numpy16_bug():
     result = ser.prod()
 
     assert not isinstance(result, Series)
+
+
+def test_sum_with_level():
+    obj = Series([10.0], index=MultiIndex.from_tuples([(2, 3)]))
+
+    result = obj.sum(level=0)
+    expected = Series([10.0], index=[2])
+    tm.assert_series_equal(result, expected)
 
 
 @pytest.mark.parametrize("func", [np.any, np.all])
